ref(tests): move unwrapSOPSData tests to separate file and remove duplicates from vault encryption tests
