+++ internal/installer/export_test.go
@@ -64,5 +64,3 @@ func (o *OpenBaoInstaller) ValidateConfig() error {
 func (o *OpenBaoInstaller) ReadinessTimeout() time.Duration {
 	return o.readinessTimeout()
 }
-
-var UnwrapSOPSData = unwrapSOPSData
