+++ internal/installer/vault_encryption_test.go
@@ -14,44 +14,6 @@ import (
 	"github.com/codesphere-cloud/oms/internal/installer"
 )
 
-var _ = Describe("unwrapSOPSData", func() {
-	It("returns data unchanged when there is no data: wrapper", func() {
-		input := []byte("secrets:\n    - name: foo\n      fields:\n        password: bar\n")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal(string(input)))
-	})
-
-	It("strips a top-level data: | wrapper and returns inner content", func() {
-		input := []byte("data: |\n    secrets:\n        - name: foo\n          fields:\n            password: bar\n")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal("secrets:\n    - name: foo\n      fields:\n        password: bar\n"))
-	})
-
-	It("returns data unchanged for an empty document", func() {
-		input := []byte("")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal(string(input)))
-	})
-
-	It("returns data unchanged when root has multiple keys", func() {
-		input := []byte("data: some-value\nsops:\n  key: val\n")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal(string(input)))
-	})
-
-	It("returns data unchanged for invalid YAML", func() {
-		input := []byte("not: valid: yaml: [[")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal(string(input)))
-	})
-
-	It("returns data unchanged when data is not a scalar", func() {
-		input := []byte("data:\n  nested: value\n")
-		output := installer.UnwrapSOPSData(input)
-		Expect(string(output)).To(Equal(string(input)))
-	})
-})
-
 func sopsAndAgeAvailable() bool {
 	if _, err := exec.LookPath("sops"); err != nil {
 		return false
